Fixing typo in component name
